% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_missing_wrap.R
\name{convert_missing_wrap}
\alias{convert_missing_wrap}
\title{Clean up missing values in UP data}
\usage{
convert_missing_wrap(x)
}
\arguments{
\item{x}{Vector containing untransformed WRAP data.}
}
\value{
Vector containing cleaned UP data.
}
\description{
This function takes in a column of raw UP data and returns a column with missing value codes replaced as NA.
}
\seealso{
\code{\link{UP_harm}}
}
\keyword{UP}
