% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emj_min.R
\name{emj_min}
\alias{emj_min}
\title{Min and max with NA}
\usage{
emj_min(x)
}
\arguments{
\item{x}{Vector to compute minimum or maximum of.}
}
\description{
These functions take in a vector of numbers and returns the minimum or maximum.
If no non-missing values are present, NA is returned.
}
