% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emj_round.R
\name{emj_round}
\alias{emj_round}
\title{Round nicely}
\usage{
emj_round(number, lb = 0, censor = FALSE)
}
\arguments{
\item{number}{Number to round.}

\item{lb}{Lower bound. An entry whose absolute value is below this limit will be handled separately.}

\item{censor}{Governs the behavior of the lb parameter. If TRUE, absolute values below the lower bound will be formatted as left-censored (e.g., "<.0001"). If FALSE, values below the lower bound will be rounded to 0. When no lower bound is specified, censor=TRUE has no effect.}
}
\description{
This function takes in a number or vector of numbers and rounds it to two decimal 
places for large numbers, or two significant figures for small ones.
It returns a string or vector of strings.
}
