% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMyVar.R
\name{findMyVar}
\alias{findMyVar}
\title{findMyVar}
\usage{
findMyVar(freeze, x)
}
\arguments{
\item{freeze}{String containing name of ODBC object defining the freeze.}

\item{x}{String text to match using grepl. Search is case-insensitive.}
}
\value{
Data frame containing potential matches.
}
\description{
This function takes in a freeze name and a string and returns all table-field
pairs in the freeze which partly match the string.
}
\keyword{WRAP}
