% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_round.R
\name{golden_round}
\alias{golden_round}
\title{Round figures nicely}
\usage{
golden_round(x, as.char = TRUE)
}
\arguments{
\item{x}{Value to round.}
}
\value{
Rounded value.
}
\description{
This function rounds an input value based on modified Roche rules.
Absolute values between 0 and 1 are rounded to two significant digits;
between 1 and 100, to two decimal places; and above 100, to the nearest 
whole number. Non-numeric values are returned as-is.
}
\keyword{emj,}
\keyword{round}
