% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nano_to_pico.R
\name{nano_to_pico}
\alias{nano_to_pico}
\title{Convert nano to pico}
\usage{
nano_to_pico(x)
}
\arguments{
\item{x}{Value to convert.}
}
\value{
Converted value.
}
\description{
This function is suitable for converting a measurement in nanograms/mL to 
picograms/mL. If the input is character, the function will coerce the input
to numeric first, then reconvert it to character before returning it.
}
\keyword{conversion}
\keyword{emj,}
