% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_lookup.R
\name{wrap_lookup}
\alias{wrap_lookup}
\title{Look up identifiers in WRAP database}
\usage{
wrap_lookup(wrap = NULL, reggie = NULL, dbid = NULL, get.vis = FALSE)
}
\arguments{
\item{wrap}{Vector containing WRAPNos.}

\item{reggie}{Vector containing Reggieids.}

\item{dbid}{Vector containing DBIDs.}
}
\value{
Data frame containing cleaned UP data.
}
\description{
This function takes in a participant ID and returns a data frame with all IDs and, optionally, visit dates.
}
\seealso{
\code{\link{UP_harm}}
}
\keyword{WRAP}
