% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emj_round.R
\name{emj_round}
\alias{emj_round}
\title{Round nicely}
\usage{
emj_round(number)
}
\arguments{
\item{number}{Number to round.}
}
\description{
This function takes in a number or vector of numbers and rounds it to two decimal 
places for large numbers, or two significant figures for small ones.
It returns a string or vector of strings.
}
